Move planner engine connection url to properties, handling errors in rest resource, get Results request

+++ src/main/java/com/teammachine/staffrostering/web/rest/PlanningJobResource.java
@@ -2,9 +2,14 @@ package com.teammachine.staffrostering.web.rest;
 
 import com.codahale.metrics.annotation.Timed;
 import com.teammachine.staffrostering.domain.PlanningJob;
+import com.teammachine.staffrostering.domain.StaffRosterParametrization;
 import com.teammachine.staffrostering.domain.enumeration.JobStatus;
 import com.teammachine.staffrostering.planner.PlannerService;
+import com.teammachine.staffrostering.planner.PlannerServiceJob;
 import com.teammachine.staffrostering.repository.PlanningJobRepository;
+import com.teammachine.staffrostering.web.rest.dto.PlanningJobWithResultDTO;
+import com.teammachine.staffrostering.web.rest.errors.CustomParameterizedException;
+import com.teammachine.staffrostering.web.rest.errors.ErrorConstants;
 import com.teammachine.staffrostering.web.rest.util.HeaderUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,7 +22,9 @@ import javax.inject.Inject;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
-import java.util.Optional;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping("/api")
@@ -26,9 +33,9 @@ public class PlanningJobResource {
     private final Logger log = LoggerFactory.getLogger(PlanningJobResource.class);
 
     @Inject
-    private PlanningJobRepository planningJobRepository;
-    @Inject
     private PlannerService plannerService;
+    @Inject
+    private PlanningJobRepository planningJobRepository;
 
     @RequestMapping(value = "/planning-jobs",
         method = RequestMethod.POST,
@@ -39,9 +46,11 @@ public class PlanningJobResource {
         if (planningJob.getId() != null) {
             return ResponseEntity.badRequest().headers(HeaderUtil.createFailureAlert("planningJob", "idexists", "A new planningJob cannot already have an ID")).body(null);
         }
-        PlanningJob job = plannerService.runPlanningJob(planningJob.getParameterization());
-        job.setParameterization(planningJob.getParameterization());
-        PlanningJob result = planningJobRepository.save(job);
+        PlannerServiceJob job = plannerService.runPlanningJob(planningJob.getParameterization())
+            .orElseThrow(() -> new CustomParameterizedException(ErrorConstants.ERR_UNABLE_TO_RUN_PLANNING_JOB));
+        planningJob.setJobId(job.getJobId());
+        planningJob.setStatus(job.getStatus());
+        PlanningJob result = planningJobRepository.save(planningJob);
         return ResponseEntity.created(new URI("/api/planning-jobs/" + result.getId()))
             .headers(HeaderUtil.createEntityCreationAlert("planningJob", result.getId().toString()))
             .body(result);
@@ -53,8 +62,29 @@ public class PlanningJobResource {
     @Timed
     public List<PlanningJob> getAllPlanningJobs() {
         log.debug("REST request to get all PlanningJobs");
-        List<PlanningJob> planningJobs = planningJobRepository.findAll();
-        return planningJobs;
+        return planningJobRepository.findAll();
+    }
+
+    @RequestMapping(value = "/planning-jobs",
+        method = RequestMethod.PUT,
+        produces = MediaType.APPLICATION_JSON_VALUE)
+    @Timed
+    public ResponseEntity<Void> syncAllPlanningJobStatuses() {
+        log.debug("REST request to sync PlanningJobs' statuses");
+        Map<String, PlannerServiceJob> allPlanningJobs = plannerService.getAllPlanningJobs().stream()
+            .collect(Collectors.toMap(PlannerServiceJob::getJobId, Function.identity()));
+        planningJobRepository.findAll().stream()
+            .forEach(persistedJob -> {
+                PlannerServiceJob job = allPlanningJobs.get(persistedJob.getJobId());
+                if (job == null) {// removed in planner engine
+                    persistedJob.setStatus(null);
+                    planningJobRepository.save(persistedJob);
+                } else if (persistedJob.getStatus() != job.getStatus()) {// status has been updated on server
+                    persistedJob.setStatus(job.getStatus());
+                    planningJobRepository.save(persistedJob);
+                }
+            });
+        return ResponseEntity.ok().build();
     }
 
     @RequestMapping(value = "/planning-jobs/{id}",
@@ -64,30 +94,39 @@ public class PlanningJobResource {
     public ResponseEntity<PlanningJob> syncPlanningJobStatus(@PathVariable Long id) {
         log.debug("REST request to sync PlanningJob status: {}", id);
         PlanningJob planningJob = planningJobRepository.findOne(id);
-        JobStatus status = plannerService.getPlanningJobStatus(planningJob);
+        if (planningJob == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        JobStatus status = plannerService.getPlanningJob(planningJob.getJobId()).map(PlannerServiceJob::getStatus).orElse(null);
         if (status != planningJob.getStatus()) {
             planningJob.setStatus(status);
             planningJobRepository.save(planningJob);
         }
-        return Optional.ofNullable(planningJob)
-            .map(result -> new ResponseEntity<>(
-                result,
-                HttpStatus.OK))
-            .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
+        return ResponseEntity.ok(planningJob);
     }
 
     @RequestMapping(value = "/planning-jobs/{id}",
         method = RequestMethod.GET,
         produces = MediaType.APPLICATION_JSON_VALUE)
     @Timed
-    public ResponseEntity<PlanningJob> getPlanningJob(@PathVariable Long id) {
+    public ResponseEntity<PlanningJobWithResultDTO> getPlanningJob(@PathVariable Long id) {
         log.debug("REST request to get PlanningJob : {}", id);
         PlanningJob planningJob = planningJobRepository.findOne(id);
-        return Optional.ofNullable(planningJob)
-            .map(result -> new ResponseEntity<>(
-                result,
-                HttpStatus.OK))
-            .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
+        if (planningJob == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        PlanningJobWithResultDTO jobDTO = new PlanningJobWithResultDTO();
+        jobDTO.setId(planningJob.getId());
+        jobDTO.setJobId(planningJob.getJobId());
+        jobDTO.setStatus(planningJob.getStatus());
+        StaffRosterParametrization parameterization = planningJob.getParameterization();
+        jobDTO.setParameterization(parameterization);
+        plannerService.getPlanningJob(planningJob.getJobId()).ifPresent(job -> {
+            jobDTO.setShiftAssignments(job.getResult().getShiftAssignments());
+            parameterization.setHardConstraintMatches(job.getResult().getHardConstraintMatches());
+            parameterization.setSoftConstraintMatches(job.getResult().getSoftConstraintMatches());
+        });
+        return ResponseEntity.ok(jobDTO);
     }
 
     @RequestMapping(value = "/planning-jobs/{id}",
@@ -97,7 +136,10 @@ public class PlanningJobResource {
     public ResponseEntity<Void> deletePlanningJob(@PathVariable Long id) {
         log.debug("REST request to delete PlanningJob : {}", id);
         PlanningJob planningJob = planningJobRepository.findOne(id);
-        plannerService.terminateAndDeleteJob(planningJob);
+        if (planningJob == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        plannerService.terminateAndDeleteJob(planningJob.getJobId());
         planningJobRepository.delete(id);
         return ResponseEntity.ok().headers(HeaderUtil.createEntityDeletionAlert("planningJob", id.toString())).build();
     }
